+++ rushstr-core/src/stores/store_trait.rs
@@ -1,0 +1,4 @@
+pub trait StoreTrait {
+    fn filter_items(&self, input: &str) -> Vec<String>;
+    fn total(&self) -> usize;
+}
