+++ rushstr-core/src/types/common_types.rs
@@ -1,0 +1,4 @@
+use crate::StoreTrait;
+use std::sync::Arc;
+
+pub type Store = Arc<dyn StoreTrait>;
