+++ rushstr-tui/Cargo.toml
@@ -4,3 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+ratatui.workspace = true
+anyhow.workspace = true
+rushstr-core = { path = "../rushstr-core" }
