+++ rushstr-tui/src/ux/mod.rs
@@ -1,0 +1,1 @@
+pub mod search_box;
